refactor(mm): separate physical address width from native pointers

+++ kernel/src/abi/linux/generic/futex.rs
@@ -35,7 +35,7 @@ enum FutexKey {
         address: usize,
     },
     Shared {
-        physical_address: usize,
+        physical_address: u64,
     },
 }
