+++ kernel/src/drivers/mmc/core.rs
@@ -453,7 +453,7 @@ impl MemoryMappingOps for EmmcBlockDevice {
         Err("Memory mapping not supported by eMMC")
     }
 
-    fn on_mapped(&self, _vaddr: usize, _paddr: usize, _length: usize, _offset: usize) {}
+    fn on_mapped(&self, _vaddr: usize, _paddr: u64, _length: usize, _offset: usize) {}
 
     fn on_unmapped(&self, _vaddr: usize, _length: usize) {}
