+++ kernel/src/fs/vfs_v2/core.rs
@@ -482,7 +482,7 @@ impl MemoryMappingOps for VfsFileObject {
         self.inner.get_mapping_info_with(offset, length, is_shared)
     }
 
-    fn on_mapped(&self, vaddr: usize, paddr: usize, length: usize, offset: usize) {
+    fn on_mapped(&self, vaddr: usize, paddr: u64, length: usize, offset: usize) {
         self.inner.on_mapped(vaddr, paddr, length, offset);
     }
