+++ kernel/src/fs/vfs_v2/drivers/fat32/mod.rs
@@ -700,7 +700,7 @@ impl Fat32FileSystem {
         &self,
         start_cluster: u32,
         page_index: u64,
-        paddr: usize,
+        paddr: u64,
     ) -> Result<(), FileSystemError> {
         use crate::environment::PAGE_SIZE;
         use crate::vm::addr::phys_to_virt;
