+++ kernel/src/object/tests/mock.rs
@@ -77,7 +77,7 @@ impl MemoryMappingOps for MockFileObject {
         Err("Memory mapping not supported")
     }
 
-    fn on_mapped(&self, _vaddr: usize, _paddr: usize, _length: usize, _offset: usize) {}
+    fn on_mapped(&self, _vaddr: usize, _paddr: u64, _length: usize, _offset: usize) {}
 
     fn on_unmapped(&self, _vaddr: usize, _length: usize) {}
 
@@ -195,7 +195,7 @@ impl MemoryMappingOps for MockTaskFileObject {
         Err("Memory mapping not supported")
     }
 
-    fn on_mapped(&self, _vaddr: usize, _paddr: usize, _length: usize, _offset: usize) {}
+    fn on_mapped(&self, _vaddr: usize, _paddr: u64, _length: usize, _offset: usize) {}
 
     fn on_unmapped(&self, _vaddr: usize, _length: usize) {}
